Add controller for Surveys
Added endpoint methods needed for the Survey model.  While not
required by the exercise, an index method is included for a convenient
view of all the submitted surveys.

The store() method includes validation, requring data depending on the
submitted values.  For instance, the daily frequency valus is only
required when the headache frequency value is daily.  The constant
values defined in the Model are used to maintain a single source of
truth for what values are permitted.

Extra methods were included for the unused routes for possible future
use.